Add Normalization of Drive Powers
Many times we want to send the robot into full power, however that may lead to unintended consequences, if the driveteam is having issues with the robot seeming to not be able to strafe while driving, or just general funkyness, I would recommend turning this on. Motors can only run to a power of 1, the gamepad gives us a value of one when at full power, meaning full power forward, there is no change when adding the power of a turn, this is seen more when you try to turn, strafe, and drive at full power. This setting will temporarily make the robot seem slower, but it will be more responsive to driver control.